Bug fixes and added comments

+++ api/views.py
@@ -1,8 +1,8 @@
+from os import stat
 from pymongo import DESCENDING
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
-from django.conf import settings
 from .models import (
     SearchQuery, UserQuery, YTVideo
 )
@@ -14,19 +14,37 @@ from bson import ObjectId
 
 
 @api_view(['POST'])
-def post_query(request):
-    
+def register_query(request):
+    '''
+    View which allows a user to register a query. Requires authentication
+
+    Request parameters:
+    - query: query which is to be searched
+
+    Response:
+    - On success
+        - returns a message that query has been stored
+    - On failure
+        - returns error message
+    '''
+
     try:
-        query = request.POST['query']
+        # check if query is valid
+        query = request.data.get('query', None)
+        if query == None:
+            return Response({"message": "Query not found"})
         if len(query)==0:
-            return Response({"Message":"Empty Query"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message":"Empty Query"}, status=status.HTTP_400_BAD_REQUEST)
         elif len(query)>100:
-            return Response({"Message":"Query length exceeded 100"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"message":"Query length exceeded 100"}, status=status.HTTP_400_BAD_REQUEST)
         
+        # extract access token from header and get user
         authorization_header = request.headers.get('Authorization')
         access_token = authorization_header.split(' ')[1]
         user = get_user_from_token(access_token)
         
+        # check if user already has query with same name
+        # if not, then create the query
         user_obj_id = ObjectId(user._id)
         query_obj = UserQuery.objects.raw({'user': user_obj_id, 'query': query})
         if query_obj.count() == 0:
@@ -37,35 +55,67 @@ def post_query(request):
             )
             user_query.save()
         
+        # fetch the videos associated with this single query
         fetch_single_video.delay(query)
-        return Response({"Message": "Query stored"}, status=status.HTTP_201_CREATED)
+        return Response({"message": "Query stored"}, status=status.HTTP_201_CREATED)
     except Exception as e:
-        return Response({"Message": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+        print(e)
+        return Response({"message": "Invalid request"}, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['GET'])
 def get_query_results(request):
+    '''
+    View which allows a user to obtain the videos associated with a query. Requires authentication
+
+    Request parameters(parameters should be present in url):
+    - query(Required): the query to be searched
+    - number(Required): number of videos to obtain(between 1 and 20)
+
+    Response:
+    - On success:
+        - query: query that was searched
+        - number: number of videos requested
+        - message(Optional): if no video is found, it returns a message
+        - videos: a list consisting of videos:
+            - video_id
+            - video_title
+            - video_description
+            - channel_title
+            - channel_id
+            - publish_time
+            - time_created
+            - time updated
+    - On failure:
+        - returns the error
+    '''
+
+    # get the parameters and validate
     try:
         query = request.query_params.get('query')
         number = int(request.query_params.get('number', 5))
     except:
-        return Response({"Message": "Invalid parameters"}, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message": "Invalid parameters"}, status=status.HTTP_400_BAD_REQUEST)
 
     if number > 20 or number<=0:
-        return Response({"Message": "Number should be between 1 and 20"}, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message": "Number should be between 1 and 20"}, status=status.HTTP_400_BAD_REQUEST)
     
+    # extract access token from header and get user
     authorization_header = request.headers.get('Authorization')
     access_token = authorization_header.split(' ')[1]
     user = get_user_from_token(access_token)
+    
+    # Search if any such query is associated with the user 
     user_object_id = ObjectId(user._id)
     if UserQuery.objects.raw({'query': query, 'user': user_object_id}).count()==0:
-        return Response({"Message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
+        return Response({"message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
 
     try:
         search_query = SearchQuery.objects.get({'query': query})
     except:
-        return Response({"Message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
+        return Response({"message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
 
+    # return the data
     data = {}
     data['query'] = query
     data['number'] = number
@@ -77,34 +127,67 @@ def get_query_results(request):
         data['videos'] = serializer.data
     except:
         data['videos'] = []
-        data['Message'] = "No video found"
+        data['message'] = "No video found"
     return Response(data, status=status.HTTP_200_OK)
 
 
 @api_view(['GET'])
 def get_user_queries(request):
+    '''
+    View which returns all the queries associated with the user. Requires authentication
+    
+    Request parameters:
+    None
+
+    Response:
+    - On success:
+        - queries: a list of strings denoting queries
+    - On failure
+        - returns the error
+    '''
+
+    # extract access token from header and get user
     authorization_header = request.headers.get('Authorization')
     access_token = authorization_header.split(' ')[1]
     user = get_user_from_token(access_token)
 
+    # fetch all the queries associated with the current user
     user_obj_id = ObjectId(user._id)
     if UserQuery.objects.raw({'user': user_obj_id}).count()>0:
         user_queries = [query.query for query in list(UserQuery.objects.raw({'user': user_obj_id}))]
         return Response({"queries": user_queries}, status=status.HTTP_200_OK)
     else:
-        return Response({"Message": "No queries associated with the user"}, status=status.HTTP_204_NO_CONTENT)
+        return Response({"message": "No queries associated with the user"}, status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['POST'])
 def delete_query(request):
+    '''
+    View which deletes the query for the current user. Requires authentication
+    
+    Request parameters
+    - query(Required): query which is to be deleted
+
+    Response:
+    - On success:
+        - returns a message that query has been deleted
+    - On failure
+        - returns the error
+    '''
+
+    # extract the access token from the header and get user
     authorization_header = request.headers.get('Authorization')
     access_token = authorization_header.split(' ')[1]
     user = get_user_from_token(access_token)
-    query = request.POST['query']
+    query = request.data.get('query', None)
+    if query == None:
+        return Response({"message": "query not found in parameters"}, status=status.HTTP_400_BAD_REQUEST)
+
+    # delete the record if it exists or return the error
     user_obj_id = ObjectId(user._id)
     try:
         user_query = UserQuery.objects.get({'user': user_obj_id, 'query': query})
         user_query.delete()
-        return Response({"Message": "Deleted query"}, status=status.HTTP_200_OK)
+        return Response({"message": "Deleted query"}, status=status.HTTP_200_OK)
     except:
-        return Response({"Message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
+        return Response({"message": "Query not found"}, status=status.HTTP_404_NOT_FOUND)
